refactor: extract SessionLibraryEntry.swift from SessionLibraryItem.swift
Splits the `SessionLibraryEntry` concrete Codable value (~78 lines)
out of SessionLibraryItem.swift so that file holds only the protocol
+ the generic snapshot value.

Byte-identical move of lines 10-88. Both files import Foundation.

SessionLibraryItem.swift: 93 → 14 lines. Tests: 667.
